+++ packages/magento-cart-checkout/index.ts
@@ -1,2 +1,5 @@
 export * from './queries/CartPage.gql'
+export * from './queries/CheckoutAgreements.gql'
 export * from './ShippingPage.gql'
+
+export * from './components'