+++ packages/magento-newsletter/components/index.ts
@@ -1,0 +1,2 @@
+export { default as CustomerNewsletterToggle } from './CustomerNewsletterToggle'
+export { default as SignupNewsletter } from './SignupNewsletter'