+++ packages/magento-cart-billing-address/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.103.0](https://github.com/ho-nl/m2-pwa/compare/@graphcommerce/magento-cart-billing-address@2.102.19...@graphcommerce/magento-cart-billing-address@2.103.0) (2021-09-24)
+
+### Bug Fixes
+
+- cart cache not up-to-date
+  ([aae4d30](https://github.com/ho-nl/m2-pwa/commit/aae4d302e320475d28a356ec304ea6afb64c3080))
+- edit billing address updating state
+  ([ecd9f48](https://github.com/ho-nl/m2-pwa/commit/ecd9f48ce313d8e7a698c06ff29b88231dc50168))
+
+### Features
+
+- edit billing address on checkout payment step
+  ([96a5719](https://github.com/ho-nl/m2-pwa/commit/96a5719437616006efb2588c3516d3f2608c1fb8))
+
+# [2.102.0](https://github.com/ho-nl/m2-pwa/compare/@graphcommerce/magento-cart-billing-address@2.101.10...@graphcommerce/magento-cart-billing-address@2.102.0) (2021-08-12)
+
 # [2.102.0](https://github.com/ho-nl/m2-pwa/compare/@graphcommerce/magento-cart-billing-address@2.101.10...@graphcommerce/magento-cart-billing-address@2.102.0) (2021-08-12)
 
 ### Features