+++ packages/magento-cart-checkout/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.103.0](https://github.com/ho-nl/m2-pwa/compare/@graphcommerce/magento-cart-checkout@2.102.20...@graphcommerce/magento-cart-checkout@2.103.0) (2021-09-24)
+
+### Bug Fixes
+
+- agreements positioning
+  ([89c2dee](https://github.com/ho-nl/m2-pwa/commit/89c2dee1debeb84c8b2cd9abaac85f03759604c8))
+- full text link
+  ([1186248](https://github.com/ho-nl/m2-pwa/commit/1186248d47898649b2f06c63a78cf254396c7f60))
+- top spacing
+  ([486fdb0](https://github.com/ho-nl/m2-pwa/commit/486fdb0536e5f6b08f293ea442661d7ad7541717))
+
+### Features
+
+- **checkout:** checkout agreements checkboxes in checkout
+  ([a8b4ddb](https://github.com/ho-nl/m2-pwa/commit/a8b4ddb3a9750c2b7ff86cd460e0ff7fc4cc0ad1))
+- **payment-agreements-form:** checkout agreements checkboxes
+  ([13c8164](https://github.com/ho-nl/m2-pwa/commit/13c816499d220a2ce940672a95beca508b78ddc9))
+- read checkout policies in sidebar sheet
+  ([3fb765e](https://github.com/ho-nl/m2-pwa/commit/3fb765e14a8cfaf0bb27acd8368926ac27ed6a4c))
+
+## [2.102.5](https://github.com/ho-nl/m2-pwa/compare/@graphcommerce/magento-cart-checkout@2.102.4...@graphcommerce/magento-cart-checkout@2.102.5) (2021-08-13)
+
 ## [2.102.5](https://github.com/ho-nl/m2-pwa/compare/@graphcommerce/magento-cart-checkout@2.102.4...@graphcommerce/magento-cart-checkout@2.102.5) (2021-08-13)
 
 ### Bug Fixes